Squashed commit of the following:
commit 058affcadae200eca51107a4049c26a522843abd
Author: cox <adam.cox@kit.edu>
Date:   Sun Oct 14 18:38:10 2012 +0200

    removing ktemplate from automatically being built. its just an example and shouldn't be part of the official build.

commit fbd2fc8df6d2300931c82f8cd34b42a8ae0e1d50
Author: cox <adam.cox@kit.edu>
Date:   Sun Oct 14 18:21:58 2012 +0200

    removing unnecessary classes and fixing all compiler warnings

    hopefully i didn't break anything in the KWindowDesign class with the casting of unsigned ints to ints. I don't think that I did… but something to watch out for.

commit ffdc20fd4e86f3ee8356810123f8a6fd80c0ca98
Author: cox <adam.cox@kit.edu>
Date:   Wed Oct 10 12:36:12 2012 +0200

    first commit on cleanup

git-svn-id: https://edwdev-ik.fzk.de/svn/EDELWEISS/analysis/kdata/trunk@1728 e4a34bab-3947-0410-866c-ab59ef520b34

+++ kpta/KPulseShifter.cxx
@@ -58,10 +58,10 @@ bool KPulseShifter::RunProcess(void)
       {
         // Shifting and filling the unknown points at the edge with the first/last values in the trace
         for(unsigned int i = 0 ; i < fOutputSize; i++)
-          if( (int)(i - fPulseShift)>=0 &&  (int)(i-fPulseShift)<(int)fOutputSize)
+          if( (int)i - fPulseShift >= 0 &&  (int)i-fPulseShift < (int)fOutputSize)
             *(fOutputPulse+i) = *(fInputPulse+i-fPulseShift);
           else{
-            if((i - fPulseShift)<0)
+            if( (int)i - fPulseShift < 0)
               *(fOutputPulse+i) = *(fInputPulse);
             else
               *(fOutputPulse+i) = *(fInputPulse+fOutputSize-1);
